room-core/requirements.txt
argcomplete==1.11.1
astroid==2.3.3
boto3==1.12.41
botocore==1.15.41
certifi==2020.4.5.1
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.1
dnspython==1.16.0
docutils==0.15.2
durationpy==0.5
exponent-server-sdk==0.3.1
Flask==1.1.2
future==0.18.2
fuzzywuzzy==0.18.0
hjson==3.0.1
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.5
kappa==0.6.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pip-tools==5.0.0
placebo==0.9.0
plaid-python==3.9.0
PyJWT==1.7.1
pylint==2.4.4
pymongo==3.10.1
python-dateutil==2.6.1
python-slugify==4.0.0
PyYAML==5.3.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
text-unidecode==1.3
toml==0.10.0
tqdm==4.45.0
troposphere==2.6.0
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==0.16.1
wrapt==1.11.2
wsgi-request-logger==0.4.6
zappa==0.51.0
zipp==3.1.0
